# docs/openapi.yaml
openapi: 3.0.0

info:
  title: tomeit API
  version: 0.1.0

servers:
  - url: http://localhost:8080/v0
    description: 開発用ローカルサーバ

paths:
  /tasks:
    post:
      tags:
        - tasks
      summary: 新しいタスクを作成する
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  example: タスク1
                expectedPomodoroNum:
                  type: integer
                  default: 0
                dueOn:
                  type: string
                  format: date-time
                  default: ""
      responses:
        "201":
          description: 成功
          headers:
            Location:
              schema:
                type: string
                example: https://api.tomeit.com/v0/tasks/1
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/AuthenticationError"
      security:
        - bearerAuth: [ ]
    get:
      tags:
        - tasks
      summary: タスク一覧を取得する
      description: 作成日時の昇順で最大30件まで取得する.
      parameters:
        - name: isCompleted
          description: タスクが完了したかどうかで絞り込みを行う。
          in: "query"
          schema:
            type: string
            format: boolean
        - name: completedOn
          description: タスクを完了した日付で絞り込む。
          in: "query"
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tasks"
        "401":
          $ref: "#/components/responses/AuthenticationError"
      security:
        - beaerAuth: [ ]
  /tasks/{taskID}:
    put:
      summary: 特定のタスクを更新する
      tags:
        - tasks
      parameters:
        - $ref: "#/components/parameters/taskID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - expectedPomodoroNum
                - dueOn
                - isCompleted
              properties:
                title:
                  type: string
                  example: タスク1
                expectedPomodoroNum:
                  type: integer
                  example: 2
                dueOn:
                  type: string
                  format: date-time
                isCompleted:
                  type: boolean
                  example: false
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/AuthenticationError"
        "403":
          $ref: "#/components/responses/AuthorizationError"
        "404":
          $ref: "#/components/responses/NotFoundError"
      security:
        - beaerAuth: [ ]
    delete:
      summary: 特定のタスクを削除する
      tags:
        - tasks
      parameters:
        - $ref: "#/components/parameters/taskID"
      responses:
        "204":
          description: 成功
        "401":
          $ref: "#/components/responses/AuthenticationError"
        "403":
          $ref: "#/components/responses/AuthorizationError"
        "404":
          $ref: "#/components/responses/NotFoundError"
      security:
        - beaerAuth: [ ]
  /pomodoros:
    post:
      summary: 新しいポモドーロ記録を作成する
      tags:
        - pomodoros
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - taskID
              properties:
                taskID:
                  type: integer
                  example: 1
      responses:
        "201":
          description: 成功
          headers:
            Location:
              schema:
                type: string
                example: https://api.tomeit.com/v0/pomodoros/1
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pomodoro"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/AuthenticationError"
      security:
        - beaerAuth: [ ]
    get:
      summary: ポモドーロ記録一覧を取得する
      tags:
        - pomodoros
      description: 作成日時の昇順で最大30件まで取得する.
      parameters:
        - name: createdOn
          in: "query"
          description: 作成日付で絞り込む
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pomodoros"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/AuthenticationError"
      security:
        - beaerAuth: [ ]
  /pomodoros/{pomodoroID}:
    delete:
      summary: 特定のポモドーロ記録を削除する
      tags:
        - pomodoros
      parameters:
        - $ref: "#/components/parameters/pomodoroID"
      responses:
        "204":
          description: 成功
        "401":
          $ref: "#/components/responses/AuthenticationError"
        "403":
          $ref: "#/components/responses/AuthorizationError"
        "404":
          $ref: "#/components/responses/NotFoundError"
      security:
        - beaerAuth: [ ]
  /pomodoros/rest-count:
    get:
      summary: 次の長い休憩までのカウントを取得する
      tags:
        - pomodoros
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  restCount:
                    type: integer
                    example: 3
        "401":
          $ref: "#/components/responses/AuthenticationError"
      security:
        - beaerAuth: [ ]
  /helth:
    get:
      tags:
        - monitoring
      summary: サーバの状態を確認する
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    default: OK

components:
  schemas:
    Task:
      type: object
      required:
        - id
        - title
        - expectedPomodoroNum
        - actualPomodoroNum
        - dueOn
        - isCompleted
        - completedOn
        - createdAt
        - updatedAt
      properties:
        id:
          description: タスク ID
          type: integer
          minimum: 1
          example: 1
        title:
          description: タイトル
          type: string
          example: タスク1
        expectedPomodoroNum:
          description: 予測ポモドーロ数
          type: integer
          minimum: 0
          maximum: 6
          default: 4
        actualPomodoroNum:
          description: 完了したポモドーロ数
          type: integer
          minimum: 0
          default: 0
        dueOn:
          description: |
            期限日付

            存在しない場合は ""（空の文字列）を返す。
          type: string
          format: date-time
        isCompleted:
          description: 完了したかどうか
          type: boolean
        completedOn:
          description: |
            完了した日付

            タスクが完了済みでない場合は ""（空の文字列）を返す。
          type: string
          format: date-time
        createdAt:
          description: 作成日時
          type: string
          format: date-time
        updatedAt:
          description: 更新日時
          type: string
          format: date-time
    Tasks:
      type: object
      required:
        - tasks
      properties:
        tasks:
          description: タスク一覧
          type: array
          items:
            $ref: "#/components/schemas/Task"
    Pomodoro:
      type: object
      required:
        - id
        - task
        - createdAt
      properties:
        id:
          description: ポモドーロ記録 ID
          type: integer
          minimum: 1
          example: 1
        task:
          $ref: "#/components/schemas/Task"
        createdAt:
          description: 作成日時
          type: string
          format: date-time
    Pomodoros:
      type: object
      required:
        - pomodoros
      properties:
        pomodoros:
          description: ポモドーロ記録一覧
          type: array
          items:
            $ref: "#/components/schemas/Pomodoro"
  responses:
    BadRequestError:
      description: リクエストが間違っている
      content:
        application/json:
          schema:
            type: object
            required:
              - status
              - error
            properties:
              status:
                description: ユーザ向けの大まかなエラーの説明
                type: string
                example: リクエストに何らかの間違いがあります。
              error:
                description: 開発者向けの具体的なエラーの説明
                type: string
                example: some error messages
    AuthenticationError:
      description: 認証が必要である
      content:
        application/json:
          schema:
            type: object
            required:
              - status
              - error
            properties:
              status:
                description: ユーザ向けの大まかなエラーの説明
                type: string
                example: ログインする必要があります。
              error:
                description: 開発者向けの具体的なエラーの説明
                type: string
                example: some error messages
    AuthorizationError:
      description: リソースへのアクセス権限がない
      content:
        application/json:
          schema:
            type: object
            required:
              - status
              - error
            properties:
              status:
                description: ユーザ向けの大まかなエラーの説明
                type: string
                example: 指定されたリソースに対してアクセス権限がありません。
              error:
                description: 開発者向けの具体的なエラーの説明
                type: string
                example: some error messages
    NotFoundError:
      description: リソースが存在しない
      content:
        application/json:
          schema:
            type: object
            required:
              - status
              - error
            properties:
              status:
                description: ユーザ向けの大まかなエラーの説明
                type: string
                example: 指定されたリソースが存在しません。
              error:
                description: 開発者向けの具体的なエラーの説明
                type: string
                example: some error messages
    InternalServerError:
      description: サーバ側で何らかのエラーが発生した
      content:
        application/json:
          schema:
            type: object
            required:
              - status
              - error
            properties:
              status:
                description: ユーザ向けの大まかなエラーの説明
                type: string
                example: サーバ側で何らかのエラーが発生しました。
              error:
                description: 開発者向けの具体的なエラーの説明
                type: string
                example: some error messages
  parameters:
    taskID:
      name: taskID
      in: path
      description: 一意のタスクを特定するタスク ID
      required: true
      schema:
        type: integer
        minimum: 1
    pomodoroID:
      name: pomodoroID
      in: path
      description: 一意のポモドーロ記録を特定するポモドーロ記録 ID
      required: true
      schema:
        type: integer
        minimum: 1
  securitySchemes:
    firebaseAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

tags:
  - name: tasks
    description: タスク関連 API
  - name: pomodoros
    description: ポモドーロ関連 API
  - name: monitoring
    description: モニタリング関連 API